Move Chatwoot sync action to contact infolist
