+++ app/Filament/Widgets/Stripe/CustomerInfolist.php
@@ -4,7 +4,6 @@ namespace App\Filament\Widgets\Stripe;
 
 use App\Filament\Widgets\BaseSchemaWidget;
 use App\Jobs\Stripe\CreateCustomerPortalSessionLink;
-use App\Jobs\Stripe\SyncCustomerFromChatwootContact;
 use App\Support\Dashboard\Concerns\InteractsWithDashboardContext;
 use Filament\Actions\Action;
 use Filament\Infolists\Components\TextEntry;
@@ -56,7 +55,6 @@ class CustomerInfolist extends BaseSchemaWidget
     public function schema(Schema $schema): Schema
     {
         $chatwootContext = $this->chatwootContext();
-        $contactReady = $chatwootContext->hasContact();
         $portalReady = $this->stripeContext()->hasCustomer()
             && $chatwootContext->accountId !== null
             && $chatwootContext->conversationId !== null
@@ -78,12 +76,6 @@ class CustomerInfolist extends BaseSchemaWidget
                             ->color($portalReady ? 'warning' : 'gray')
                             ->disabled(! $portalReady)
                             ->action(fn () => $this->sendCustomerPortalLink()),
-                        Action::make('fetchFromContact')
-                            ->icon(Heroicon::OutlinedArrowDownOnSquareStack)
-                            ->outlined()
-                            ->color($contactReady ? 'primary' : 'gray')
-                            ->disabled(! $contactReady)
-                            ->action(fn () => $this->syncCustomerFromChatwootContact()),
                         Action::make('reset')
                             ->action(fn () => $this->reset())
                             ->hiddenLabel()
@@ -115,53 +107,6 @@ class CustomerInfolist extends BaseSchemaWidget
                             ->placeholder('No currency'),
                     ]),
             ]);
-    }
-
-    protected function syncCustomerFromChatwootContact(): void
-    {
-        $stripeContext = $this->stripeContext();
-        $chatwootContext = $this->chatwootContext();
-
-        $customerId = $stripeContext->customerId;
-        $accountId = $chatwootContext->accountId;
-        $contactId = $chatwootContext->contactId;
-        $impersonatorId = $chatwootContext->currentUserId;
-
-        if (! $customerId) {
-            Notification::make()
-                ->title('Missing Stripe context')
-                ->body('We could not find the Stripe customer to update. Please select a customer first.')
-                ->danger()
-                ->send();
-
-            return;
-        }
-
-        if (! $accountId || ! $contactId || ! $impersonatorId) {
-            Notification::make()
-                ->title('Missing Chatwoot context')
-                ->body('We could not find the Chatwoot contact details. Please open this widget from a Chatwoot conversation.')
-                ->danger()
-                ->send();
-
-            return;
-        }
-
-        SyncCustomerFromChatwootContact::dispatch(
-            accountId: $accountId,
-            contactId: $contactId,
-            impersonatorId: $impersonatorId,
-            customerId: $customerId,
-            notifiableId: auth()->id(),
-        );
-
-        Notification::make()
-            ->title('Syncing customer details')
-            ->body('We are fetching the Chatwoot contact details and updating the Stripe customer.')
-            ->info()
-            ->send();
-
-        $this->reset();
     }
 
     protected function sendCustomerPortalLink(): void
